Corrigir fluxo de navegacao do roteiro
