--- macroChallengeApp/macroChallengeApp/Controllers/NewRoadmap/Controllers/DaysViewController.swift
@@ -1,29 +1,0 @@
-//
-//  DaysViewController.swift
-//  macroChallengeApp
-//
-//  Created by Beatriz Duque on 15/09/22.
-//
-
-import UIKit
-
-class DaysViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
