--- macroChallengeApp/macroChallengeApp/Controllers/NewRoadmap/Controllers/DestinyViewController.swift
@@ -1,29 +1,0 @@
-//
-//  DestinyViewController.swift
-//  macroChallengeApp
-//
-//  Created by Beatriz Duque on 15/09/22.
-//
-
-import UIKit
-
-class DestinyViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
